+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKGT4G8C1N9HAYSD8WY3Y3WX"
+        "01KKGZ35ZBFGRGMVB2B4AJCJP4"
     )
     r"""The unique identifier of the data source"""