+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKGT4G8BRS50JV31AJ0RCCAJ"
+        "01KKGZ35ZA67JEEP6VJPZEY6RZ"
     )
     r"""The unique identifier of the data source"""