+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-12T10:40:16.959Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-12T12:06:57.038Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")