+++ packages/orq-rc/RELEASES.md
@@ -6519,3 +6519,13 @@ Based on:
 - [python v4.6.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v4.6.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.6.0-rc.5 - packages/orq-rc
+
+## 2026-03-12 12:09:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.752.0 (2.859.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.6.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v4.6.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.6.0-rc.6 - packages/orq-rc